Update CheckoutMain default export to include setup intent provider
